+++ e2e/selectors/Receive/RequestPaymentBottomSheet.selectors.js
@@ -1,0 +1,3 @@
+export const RequestPaymentBottomSheetSelectorsIDs = {
+  REQUEST_BUTTON: 'request-payment-button',
+};
