--- e2e/selectors/Receive/RequestPaymentModal.selectors.js
@@ -1,6 +1,0 @@
-export const RequestPaymentModalSelectorsIDs = {
-  CONTAINER: 'receive-request-screen',
-  REQUEST_BUTTON: 'request-payment-button',
-  QR_MODAL: 'qr-modal',
-  ACCOUNT_ADDRESS: 'account-address',
-};
